requirements.txt
alembic==1.0.2
autopep8==1.4
certifi==2018.10.15
chardet==3.0.4
Click==7.0
flake8==3.5.0
Flask==1.0.2
Flask-Migrate==2.3.0
Flask-SQLAlchemy==2.3.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
PySocks==1.6.8
pyTelegramBotAPI==3.6.6
python-dateutil==2.7.5
python-editor==1.0.3
PyYAML==3.13
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.13
urllib3==1.23
Werkzeug==0.14.1
